+++ Dev/Dev2.Runtime.Services/Interfaces/IResourceCatalog.cs
@@ -1,8 +1,20 @@
+/*
+*  Warewolf - Once bitten, there's no going back
+*  Copyright 2021 by Warewolf Ltd <alpha@warewolf.io>
+*  Licensed under GNU Affero General Public License 3.0 or later.
+*  Some rights reserved.
+*  Visit our website for more information <http://warewolf.io/>
+*  AUTHORS <http://warewolf.io/authors.php> , CONTRIBUTORS <http://warewolf.io/contributors.php>
+*  @license GNU Affero General Public License <http://www.gnu.org/licenses/agpl-3.0.html>
+*/
+
 #pragma warning disable
+
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Text;
+using Dev2.Common;
 using Dev2.Common.Interfaces.Data;
 using Dev2.DynamicServices;
 using Dev2.Runtime.Hosting;
@@ -42,6 +54,7 @@ namespace Dev2.Runtime.Interfaces
         int GetLatestVersionNumberForResource(Guid resourceId);
         IContextualResourceCatalog NewContextualResourceCatalog(IAuthorizationService authService, Guid workspaceId);
         IWarewolfWorkflow GetWorkflow(Guid resourceId);
+        ResourceCatalogResult SaveResources(Guid serverWorkspaceID, List<DuplicateResourceTO> resourceMaps, bool overrideExisting);
     }
 
     public interface IResourceCatalogFactory